+++ CHANGELOG.md
@@ -4,6 +4,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ### Features
 
+* Add username to program count metrics
 * Log `hostid` in BOINC mode
 
 ### Enhancements